aplite-backend/requirements-backend.txt
fastapi==0.115.5
uvicorn==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
cryptography==43.0.3
python-multipart==0.0.12
sendgrid==6.11.0
